removing ag_news dep
