+++ tests/start/test_start_pt.py
@@ -1,7 +1,6 @@
 from typing import Callable, Generator
 from unittest.mock import MagicMock, patch
 
-import pandas as pd
 import torch
 import vaex
 from torch import nn
@@ -9,7 +8,6 @@ from torch.utils.data import DataLoader
 from torch.utils.data.dataset import random_split
 from torchtext.data.functional import to_map_style_dataset
 from torchtext.data.utils import get_tokenizer
-from torchtext.datasets import AG_NEWS
 from torchtext.vocab import build_vocab_from_iterator
 
 import dataquality as dq
@@ -20,10 +18,13 @@ from dataquality.schemas.task_type import TaskType
 from dataquality.utils.thread_pool import ThreadPoolManager
 from dataquality.utils.vaex import validate_unique_ids
 from tests.conftest import TestSessionVariables
+from tests.test_utils.mock_data import MockDataset, test_df, train_df
 
-train_iter = iter(AG_NEWS(split="train"))
+# Step 5: Create Dataset
+train_iter = MockDataset(train_df)
+test_iter = MockDataset(test_df)
+labels = MockDataset.labels
 tokenizer = get_tokenizer("basic_english")
-train_iter = AG_NEWS(split="train")
 
 
 def yield_tokens(data_iter):
@@ -66,7 +67,6 @@ def collate_batch(batch):
     return label_list.to(device), text_list.to(device), offsets.to(device)
 
 
-train_iter = AG_NEWS(split="train")
 dataloader = DataLoader(
     train_iter, batch_size=8, shuffle=False, collate_fn=collate_batch
 )
@@ -92,7 +92,6 @@ class TextClassificationModel(nn.Module):
         return self.classifier(embedded)
 
 
-train_iter = AG_NEWS(split="train")
 num_class = len(set([label for (label, text) in train_iter]))
 vocab_size = len(vocab)
 emsize = 64
@@ -153,22 +152,12 @@ criterion = torch.nn.CrossEntropyLoss()
 optimizer = torch.optim.SGD(model.parameters(), lr=LR)
 scheduler = torch.optim.lr_scheduler.StepLR(optimizer, 1.0, gamma=0.1)
 total_accu = None
-train_iter, test_iter = AG_NEWS()
+
 train_dataset = to_map_style_dataset(train_iter)
 num_train = int(len(train_dataset) * 0.95)
 split_train_, split_valid_ = random_split(
     train_dataset, [num_train, len(train_dataset) - num_train]
 )
-
-ag_train = to_map_style_dataset(AG_NEWS(split="train"))[:192]
-ag_test = to_map_style_dataset(AG_NEWS(split="test"))[192:220]
-train_df = pd.DataFrame(ag_train)
-test_df = pd.DataFrame(ag_test)
-train_df = train_df.reset_index().rename(columns={0: "label", 1: "text", "index": "id"})
-train_df["id"] = train_df["id"] + 10000
-test_df = test_df.reset_index().rename(columns={0: "label", 1: "text", "index": "id"})
-labels = pd.concat([train_df["label"], test_df["label"]]).unique()
-labels.sort()
 
 
 @patch.object(dq.core.init.ApiClient, "valid_current_user", return_value=True)
@@ -219,13 +208,13 @@ def test_text_pt(
     set_test_config(default_task_type=TaskType.text_classification)
 
     train_dataloader_dq = DataLoader(
-        ag_train,
+        train_iter,
         batch_size=BATCH_SIZE,
         shuffle=True,
         collate_fn=collate_batch,
     )
     test_dataloader_dq = DataLoader(
-        ag_test, batch_size=BATCH_SIZE, shuffle=True, collate_fn=collate_batch
+        test_iter, batch_size=BATCH_SIZE, shuffle=True, collate_fn=collate_batch
     )
     with DataQuality(
         modeldq,
